very large commit featuring new DB classes and set arrangement logic.
This commit include a complete rewrite of all the classes that derive
from plateDB and mangaDB. Now the Totoro classes inherit the
ModelClasses and extend them with specific methods and properties.

Also, the set arrangement logic has changed. There are now two modes
of set rearrangement. Plate.update() (which is executed on plate
loading) will look for valid exposures not assigned to a set and
append them in the best possible incomplete set or in a new one. No
rearrangement of sets is made. Plate.rearrangeSets() will perform a
complete (and slower) set rearrangement by checking every possible
permutation of exposures. This is only meant to be done before or
after observations.

minor fixes to previous lrge commit.